wcc/api/requirements.txt
arrow==0.13.1
astroid==2.2.5
autopep8==1.4.3
bson==0.5.8
Click==7.0
colorama==0.4.1
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Flask==1.0.2
geos==0.2.1
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.2
psycopg2==2.8.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.9.0b1.dev0
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
Werkzeug==0.15.2
wrapt==1.11.1
